Updated 101 to react to new attacks
We changed the way IDS101 behaves. Now it does not disconnect the sensor LIT101 on an attack detection. During an attack, IDS101 will estimate the value of LTI101 and report it to PLC101, but also continues to receive values from LIT101. Estimated and received values are compared and when and attack is no longer detected, IDS101 notifies PLC101. In that case, PLC101 continues to receive values from LIT101